requirements.txt
torch>=2.2.0
transformers>=4.57.0
datasets>=2.18.0
torchmetrics>=1.4.0
einops>=0.8.0
pyyaml>=6.0
safetensors>=0.4.0
numpy>=1.26.0
